requirements.txt
matplotlib
numpy
py-vox-io